refactor: make homepage components structure simpler
