+++ app/_features/home/home.tsx
@@ -1,6 +1,6 @@
 import Header from '@/_shared/components/header/header';
-import CreateRoom from '@/_features/home/create-room/create-room';
-import JoinRoom from '@/_features/home/join-room/join-room';
+import CreateRoom from '@/_features/home/create-room';
+import JoinRoom from '@/_features/home/join-room';
 import Footer from '@/_shared/components/footer/footer';
 
 export default function View() {
