--- app/_features/home/join-room/index.ts
@@ -1,1 +1,0 @@
-export { default } from '@/_features/home/join-room/join-room';
