--- app/_features/home/join-room/join-room.tsx
@@ -1,13 +1,0 @@
-import JoinField from '@/_features/home/join-room/join-field';
-
-export default function JoinRoom() {
-  return (
-    <form className="w-full max-w-xl rounded-md bg-zinc-800 p-6 shadow-md">
-      <h3 className="text-xl font-semibold tracking-wide">Join a room</h3>
-      <p className="mt-2 text-base text-zinc-400">
-        Have a room code? Enter the code below in order to join to other rooms.
-      </p>
-      <JoinField />
-    </form>
-  );
-}
