--- app/_features/home/view/index.ts
@@ -1,1 +1,0 @@
-export { default } from '@/_features/home/view/view';
